GJKSolver: remove malloc in EPA
This whole rewrite of GJKSolver is motivated by the fact that when using
an instance of GJKSolver multiple times, it mallocs data for EPA each time.
Note: if working with meshes, then a malloc also occurs in GJK (in the
support function).
This is not ideal for objects like `ComputeCollision` and `ComputeDistance`,
which are functors defined per single pair of collision, and which both
contain a `GJKSolver`.
(These functors are meant to be re-used so as their internal GJKSolver.)

This commit fixes this malloc issue by doing the following:
- It adds an instance of `GJK` and an instance of `EPA` in the
`GJKSolver` class.
- It leverages the newly introduced `reset` method of EPA to allocate
memory **only when needed**.
This last point is important because some functions of the `GJKSolver`
class have specializations which don't use EPA (and/or GJK), so we don't
want to allocate memory in these cases.

In the future, to lower the memory footprint of the `GJKSolver` class,
we could switch to a `data` model like in Pinocchio.
That way, `GJK` and `EPA` would only be thin interfaces implementing
the algorithms.
They would be working on a `data` object that allocated by the `GJKSolver`
when needed.